refactored CacheBackedEmbeddings, storage
