+++ libs/core/langchain_core/embeddings/__init__.py
@@ -1,0 +1,7 @@
+from langchain_core.embeddings.base import Embeddings
+from langchain_core.embeddings.cache import CacheBackedEmbeddings
+
+__all__ = [
+    "CacheBackedEmbeddings",
+    "Embeddings",
+]
