+++ libs/core/langchain_core/embeddings/base.py
