+++ libs/langchain/langchain/storage/exceptions.py
@@ -1,3 +1,3 @@
-from langchain_community.storage.exceptions import InvalidKeyException
+from langchain_core.storage.exceptions import InvalidKeyException
 
 __all__ = ["InvalidKeyException"]
